# backend/src/main/java/com/texttosql/backend/dto/auth/ForgotPasswordRequest.java
package com.texttosql.backend.dto.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record ForgotPasswordRequest(
        @NotBlank(message = "Email is required")
        @Email(message = "Email must be valid")
        String email
) {
}