# frontend/texttosql/src/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"Natural Language to SQL Conversion Using LLMs","description":"This API is designed to translate natural language queries into structured SQL commands using AI-powered Large Language Models. It aims to simplify data access for non-technical users by bridging the gap between natural language and database queries.","contact":{"name":"Ramazan Bozkurt & Abdulkadir Sözgen","email":"ramazan.bozkurt.dev@gmail.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"2. Schema Management","description":"Database schema, table, and column operations."},{"name":"4. Search","description":"Search functionality for schemas and chat history."},{"name":"3. Chat Bot","description":"Chat interaction and message history management."},{"name":"1. Authentication","description":"User registration and authentication operations."}],"paths":{"/api/v1/auth/profile":{"put":{"tags":["1. Authentication"],"operationId":"updateProfile","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["1. Authentication"],"operationId":"deleteAccount","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/schemas/import":{"post":{"tags":["2. Schema Management"],"operationId":"importSchema","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaImportRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DatabaseDto"}}}}}}},"/api/v1/schemas/databases":{"get":{"tags":["2. Schema Management"],"operationId":"getDatabases","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelDatabaseDto"}}}}}},"post":{"tags":["2. Schema Management"],"operationId":"createDatabase","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatabaseDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DatabaseDto"}}}}}}},"/api/v1/schemas/databases/{databaseId}/tables":{"get":{"tags":["2. Schema Management"],"operationId":"getTables","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TableDto"}}}}}}},"post":{"tags":["2. Schema Management"],"operationId":"createTable","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TableDto"}}}}}}},"/api/v1/schemas/databases/{databaseId}/tables/{tableId}/columns":{"get":{"tags":["2. Schema Management"],"operationId":"getColumns","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ColumnDto"}}}}}}},"post":{"tags":["2. Schema Management"],"operationId":"createColumn","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ColumnDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ColumnDto"}}}}}}},"/api/v1/chatbot/chats":{"get":{"tags":["3. Chat Bot"],"operationId":"getChats","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelChatDto"}}}}}},"post":{"tags":["3. Chat Bot"],"operationId":"createChat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatDto"}}}}}}},"/api/v1/chatbot/chats/{chatID}/messages":{"get":{"tags":["3. Chat Bot"],"operationId":"getMessages","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelMessageDto"}}}}}},"post":{"tags":["3. Chat Bot"],"operationId":"createMessage","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDto"}}}}}}},"/api/v1/auth/verify-email":{"post":{"tags":["1. Authentication"],"operationId":"verifyEmail","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/reset-password":{"post":{"tags":["1. Authentication"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/resend-verification":{"post":{"tags":["1. Authentication"],"operationId":"resendVerification","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendVerificationRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/register":{"post":{"tags":["1. Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/login":{"post":{"tags":["1. Authentication"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/auth/forgot-password":{"post":{"tags":["1. Authentication"],"operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/change-password":{"post":{"tags":["1. Authentication"],"operationId":"changePassword","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/schemas/databases/{databaseId}":{"get":{"tags":["2. Schema Management"],"operationId":"getDatabase","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DatabaseDto"}}}}}},"delete":{"tags":["2. Schema Management"],"operationId":"deleteDatabase","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["2. Schema Management"],"operationId":"updateDatabase","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatabaseDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DatabaseDto"}}}}}}},"/api/v1/schemas/databases/{databaseId}/tables/{tableId}":{"get":{"tags":["2. Schema Management"],"operationId":"getTable","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TableDto"}}}}}},"delete":{"tags":["2. Schema Management"],"operationId":"deleteTable","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["2. Schema Management"],"operationId":"updateTable","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TableDto"}}}}}}},"/api/v1/schemas/databases/{databaseId}/tables/{tableId}/columns/{columnId}":{"get":{"tags":["2. Schema Management"],"operationId":"getColumn","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"columnId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ColumnDto"}}}}}},"delete":{"tags":["2. Schema Management"],"operationId":"deleteColumn","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"columnId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["2. Schema Management"],"operationId":"updateColumn","parameters":[{"name":"databaseId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"tableId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"columnId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ColumnDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ColumnDto"}}}}}}},"/api/v1/chatbot/chats/{chatID}":{"get":{"tags":["3. Chat Bot"],"operationId":"getChat","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatDto"}}}}}},"delete":{"tags":["3. Chat Bot"],"operationId":"deleteChat","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["3. Chat Bot"],"operationId":"updateChat","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatDto"}}}}}}},"/api/v1/chatbot/chats/{chatID}/messages/{messageID}":{"delete":{"tags":["3. Chat Bot"],"operationId":"deleteMessage","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"messageID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["3. Chat Bot"],"operationId":"updateMessageContent","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"messageID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDto"}}}}}}},"/api/v1/chatbot/chats/{chatID}/messages/{messageID}/feedback":{"patch":{"tags":["3. Chat Bot"],"operationId":"updateMessageFeedback","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"messageID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDto"}}}}}}},"/api/v1/search/schema":{"get":{"tags":["4. Search"],"operationId":"searchSchema","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchemaSearchResponse"}}}}}}},"/api/v1/search/chatbot":{"get":{"tags":["4. Search"],"operationId":"searchChat","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatSearchResponse"}}}}}}},"/api/v1/chatbot/chats/{chatID}/export/markdown":{"get":{"tags":["3. Chat Bot"],"operationId":"exportChatToMarkdown","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/chatbot/chats/{chatID}/export/json":{"get":{"tags":["3. Chat Bot"],"operationId":"exportChatToJson","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/chatbot/chats/{chatID}/export/csv":{"get":{"tags":["3. Chat Bot"],"operationId":"exportChatToCsv","parameters":[{"name":"chatID","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/validate":{"get":{"tags":["1. Authentication"],"operationId":"validate","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["1. Authentication"],"operationId":"getMe","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/auth/health":{"get":{"tags":["1. Authentication"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateProfileRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":3,"pattern":"^[a-zA-Z0-9]+$"}},"required":["username"]},"SchemaImportRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":64,"minLength":2,"pattern":"^[a-zA-Z0-9_-]*$"},"description":{"type":"string","maxLength":255,"minLength":0},"jsonContent":{"type":"array","items":{"type":"object","additionalProperties":{}},"minItems":1}},"required":["jsonContent","name"]},"DatabaseDto":{"type":"object","properties":{"databaseId":{"type":"string","format":"uuid"},"name":{"type":"string","maxLength":64,"minLength":2,"pattern":"^[a-zA-Z0-9_-]*$"},"description":{"type":"string","maxLength":255,"minLength":0},"createdAt":{"type":"string","format":"date-time"}},"required":["name"]},"TableDto":{"type":"object","properties":{"tableId":{"type":"string","format":"uuid"},"databaseId":{"type":"string","format":"uuid"},"name":{"type":"string","maxLength":64,"minLength":2,"pattern":"^[a-zA-Z0-9_-]*$"},"description":{"type":"string","maxLength":255,"minLength":0},"createdAt":{"type":"string","format":"date-time"}},"required":["name"]},"ColumnDto":{"type":"object","properties":{"columnId":{"type":"string","format":"uuid"},"tableId":{"type":"string","format":"uuid"},"databaseId":{"type":"string","format":"uuid"},"name":{"type":"string","maxLength":50,"minLength":1,"pattern":"^[a-zA-Z0-9_-]*$"},"dataType":{"type":"string","maxLength":50,"minLength":0},"createdAt":{"type":"string","format":"date-time"},"primaryKey":{"type":"boolean"}},"required":["dataType","name"]},"ChatDto":{"type":"object","properties":{"chatId":{"type":"string","format":"uuid"},"name":{"type":"string","maxLength":50,"minLength":1,"pattern":"^[a-zA-Z0-9_ -.]*$"},"createdAt":{"type":"string","format":"date-time"}},"required":["name"]},"MessageDto":{"type":"object","properties":{"messageId":{"type":"string","format":"uuid"},"chatId":{"type":"string","format":"uuid"},"databaseId":{"type":"string","format":"uuid"},"content":{"type":"string","maxLength":5000,"minLength":0},"confidence":{"type":"number","format":"double"},"senderType":{"type":"string","enum":["LLM","USER"]},"feedback":{"type":"string","enum":["GOOD","BAD","NONE"]},"createdAt":{"type":"string","format":"date-time"}},"required":["content"]},"ResetPasswordRequest":{"type":"object","properties":{"token":{"type":"string","minLength":1},"newPassword":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["newPassword","token"]},"ResendVerificationRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"username":{"type":"string","maxLength":20,"minLength":3,"pattern":"^[a-zA-Z0-9]+$"},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","password","username"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"ChangePasswordRequest":{"type":"object","properties":{"newPassword":{"type":"string","maxLength":2147483647,"minLength":6},"currentPassword":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["currentPassword","newPassword"]},"FeedbackRequest":{"type":"object","properties":{"feedback":{"type":"string","enum":["GOOD","BAD","NONE"]}}},"SchemaSearchResponse":{"type":"object","properties":{"databases":{"type":"array","items":{"$ref":"#/components/schemas/DatabaseDto"}},"tables":{"type":"array","items":{"$ref":"#/components/schemas/TableDto"}},"columns":{"type":"array","items":{"$ref":"#/components/schemas/ColumnDto"}}}},"ChatSearchResponse":{"type":"object","properties":{"chats":{"type":"array","items":{"$ref":"#/components/schemas/ChatDto"}},"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessageDto"}}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModelDatabaseDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/DatabaseDto"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"PagedModelChatDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ChatDto"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"PagedModelMessageDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/MessageDto"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"UserDto":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":3,"pattern":"^[a-zA-Z0-9]+$"},"email":{"type":"string","format":"email","minLength":1},"role":{"type":"string","enum":["ADMIN","USER","UNKNOWN"]}},"required":["email","username"]}}}}